requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
sseclient-py==1.8.0
